requirements.txt
asgiref==3.5.2
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
django==4.1.4
django-environ==0.9.0
djangorestframework==3.14.0
drf-spectacular==0.25.1
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
psycopg2-binary==2.9.5
pyrsistent==0.19.2
pytz==2022.7
pyyaml==6.0
requests==2.28.1
requests-futures==1.0.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
